Refactor constant field
